+++ src/eval_framework/tasks/base.py
@@ -41,7 +41,10 @@ class Language(Enum):
     POL = "Polish"
     RUS = "Russian"
     UKR = "Ukrainian"
+
+    # additional langues with problematic codes
     HRV = "Croatian"
+    SRP = "Serbian"
 
     @classmethod
     def add_members(cls, new_members: dict[str, Any]) -> type["Language"]: