+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 - 2025-08-27: fix loading of extra task modules (skip non-evaluation BaseTasks with no NAME attribute), add test that no task with same names get registered
 - 2025-08-27: fix ZERO_SCROLLS_QMSUM missing from task_names.py
 - 2025-09-02: add pre-commit install to the docs
-- 2025-09-02: fix inconsistent language code for Croatian in INCLUDE task
+- 2025-09-02: fix inconsistent language code for Croatian/Serbian in INCLUDE task
 
 ## [0.1.0] - 2025-08-18